// apps/spruce/src/pages/projectAndRepoSettings/shared/tabs/PluginsTab/types.ts
import { ProjectType } from "../utils";

export interface PluginsFormState {
  performanceSettings: {
    perfEnabled: boolean | null;
  };
  buildBaronSettings: {
    useBuildBaron: boolean | null;
    ticketSearchProjects: Array<{ searchProject: string }>;
    ticketCreateProject: {
      createProject: string;
    };
    ticketCreateIssueType: {
      issueType: string;
    };
    fileTicketWebhook: {
      endpoint: string;
      secret: string;
    };
  };
  externalLinks: Array<{
    displayTitle: string;
    requesters: string[];
    displayName: string;
    urlTemplate: string;
  }>;
}

export type TabProps = {
  projectData?: PluginsFormState;
  projectType: ProjectType;
  repoData?: PluginsFormState;
};